add Delaunay adaptative edge trimming option
Having a fixed percentile of distances above which we discard edges is
not optimal as artifcts happen mostly on the border of the networks,
and the proportion of berder edges decreases with networks size.

I have also integrated OpenCV `dilate` function for the contacting are
method in order to speed it up dramatically.