add parallel version of build_contacting method
The build_contacting  method execution time was limited by the boolean
masking for each object (or cell), this procedure was spending too much
time looking at all pixels values that could equal the id of objects.
One solution is to split the segmentation image into several tiles,
run the build_contacting method on each tile and merge the detected
edges.
Here we use Dask to run tasks in parallel. First the optimal number
of splits is determined from the number of cores and the size of
the segmentation image (better method are possible, it may require
some benchmarking). Then in parallel tiles are extracted from the image,
pairs of nodes are computed, and then merged.